fix(Configurations) : add play,end option variables and getters
